requirements.txt
sentence-transformers
scikit-learn
langchain-community
langchain-openai
streamlit
matplotlib
plotly